+++ Stripe/StripeiOS/Source/STPPaymentMethod+BasicUI.swift
@@ -69,7 +69,8 @@ extension STPPaymentMethod: STPPaymentOption {
             .bacsDebit, .SEPADebit, .iDEAL, .FPX, .cardPresent, .giropay, .EPS, .payPal,
             .przelewy24, .bancontact,
             .OXXO, .sofort, .grabPay, .netBanking, .UPI, .afterpayClearpay, .blik,
-            .weChatPay, .boleto, .klarna, .linkInstantDebit, .affirm, .cashApp, .paynow, .zip, .revolutPay, .amazonPay, .alma, .mobilePay, .konbini, .promptPay,
+            .weChatPay, .boleto, .klarna, .linkInstantDebit, .affirm, .cashApp, .paynow, .zip, .revolutPay, .amazonPay,
+            .alma, .mobilePay, .konbini, .promptPay, .swish,
             .unknown:
             return false