+++ Stripe/StripeiOS/Source/STPPaymentMethodParams+BasicUI.swift
@@ -38,7 +38,8 @@ extension STPPaymentMethodParams: STPPaymentOption {
             return true
         case .alipay, .AUBECSDebit, .bacsDebit, .SEPADebit, .iDEAL, .FPX, .cardPresent, .giropay,
             .grabPay, .EPS, .przelewy24, .bancontact, .netBanking, .OXXO, .payPal, .sofort, .UPI,
-            .afterpayClearpay, .blik, .weChatPay, .boleto, .klarna, .linkInstantDebit, .affirm, .cashApp, .paynow, .zip, .revolutPay, .amazonPay, .alma, .mobilePay, .konbini, .promptPay,
+            .afterpayClearpay, .blik, .weChatPay, .boleto, .klarna, .linkInstantDebit, .affirm, .cashApp, .paynow,
+            .zip, .revolutPay, .amazonPay, .alma, .mobilePay, .konbini, .promptPay, .swish,
             .unknown:
             return false
         @unknown default: